Update guide-to-dataflows-for-mapping-data-flow-users.md
Moving the Replace values transform to accompany Derived Columns. This is correcting the assertion that this was an analog for Alter Row.  The documentation(https://learn.microsoft.com/en-us/azure/data-factory/data-flow-alter-row) for Alter Row clearly states 
"Use the Alter Row transformation to set insert, delete, update, and upsert policies on rows."
From the replace values documentation (https://learn.microsoft.com/en-us/power-query/replace-values) we see this is not the same "With Power Query, you can replace one value with another value wherever that value is found in a column. "