Update security-private-links-use.md
The current heading "Create a Fabric resource in the Azure portal" is misleading and causes confusion when following these instructions.  Initially I ignored this section since I had already created a 'Fabric' resource in my Azure subscription.  What this section is really instructing is to "Create a Microsoft.PowerBI private link services for Power BI resource in the Azure Portal" which is used to support creation of an Azure Private Endpoint for the Fabric resource.  Without completing this step the Private Endpoint creation form in the Azure Portal is incomplete and Azure will not let you create the PE for Fabric.