Fixing the ARM template sample as it is not a valid template.
I removed the "id" property as this is not a valid property for this resource type. Additionally I parameterized the template for ease of use. Used template functions to reduce the number of properties that need to be set. And parameterized the SKU as it was hard coded to RA-GRS, which would change the replication configuration of someone's storage account if they weren't paying close attention.