Update lakehouse-schemas.md
The code in the cross-workspace Spark SQL queries section is not representative of a cross-workspace query. It is simply joining two tables that are in different lakehouses in the same workspace. I propose adding a workspace, like "global" in the edit to the corporate.company.departments table so it would be joining cross-workspace lakehouses.