Update protect-inbound-traffic.md
On line 74, it implies that after enabling Private Link and Private Endpoint, Fabric cannot be access using the public internet. But I feel it will still be accessible unless the tenant level setting to Block Internet access is not enabled. But once the PL/PE is enabled, devices in that network automatically using the PE. Hence made some subtle changes