Update generate-unique-identifiers.md Explanation that this is a bad workaround
I think it's important that users know that this is a horrible workaround -- and it may certainly fail if the code is not wrapped in a transaction (assuming transactions are supported in Fabric!).  I hope that Microsoft doesn't think this poor "workaround" will delay the need for proper support of Identity columns that are defined as a column attribute.  That's the only way that two or more users or jobs can properly use an auto-increment column.  This workaround is a bad idea.