clarified the `exit` method throws an exception
It was immediately clear in the `exit` method's documentation that it throws an exception. originally it mentioned it returns a void. 

This commit updates the docs to make that clearer, and also updates the "exit a notebook" section by instructing the reader to **not enclose the method around a try/catch block**